Fixup: 32bit n_caches and magic
In reply to:
Why is n_caches uint64?  When reading the code, I would assume that this means it's very big.  Why not just uint ?
I can understand you want the header structure aligned, but that is independent of the user interface that is using 'n'
alos, please replace 'n' as a parameter to cache_slot or something more readable
Another thing to think about is to replace argument checking to use assert instead of return...
if (n >= dir->header->n_caches)
   return -1;