MDEV-18851: mysys copy win32 large page alloc/free from innodb
To preserve the same allocation behavior on Windows we enable
HAVE_LARGE_PAGE_OPTION for Windows.

Per review by Vlad we check that we're not passing a nNULL pointer to
VirtualFree as its handled as an error. This avoids what has been
observed in InnoDB as printing both "VirtualAlloc failed with
ERROR_INSUFFICIENT_RESOURCES" followed by "VirtualFree failed with
ERROR_INVALID_PARAMETER" on allocation errors.