Remove --basedir arg from systemd startup scripts
The arg was introduced as part of 75bcf1f9ad4a4bb5fa9cb8818abe6ace52e53d85
to fix a SELinux problem caused by mysqld_safe accessing files it should
not be via the my_which function.

The root cause for this was fixed in 10.3, via
355ee6877beccf27bcd07d845c985fc1a2845aca which eliminated the my_which
function from mysqld_safe entirely. Thus, in 10.3, this --basedir flag
is not necessary.