MDEV-20407 mysqld got signal 11; rowid filter
If a joined table is accessed by a full text index then now no range rowid
filter can be used when accessing the records of this table.

This bug was fixed by the patch for mdev-20056.
This commit adds only a test case with a query using a full text index
and a range condition for another index.