MDEV-21383: Possible range plan is not used under certain conditions
(10.3's variant of the fix)

make_join_select() has a section of code which starts with
 "We plan to scan all rows. Check again if we should use an index."

the code in that section will [unnecessarily] re-run the range
optimizer using this condition:

  condition_attached_to_current_table AND current_table's_ON_expr

Note that the original invocation of range optimizer in
make_join_statistics was done using the whole select's WHERE condition.

taking the whole select's WHERE condition and using multiple-equalities
allowed the range optimizer to do more inferences.

This fix uses a very conservative approach: we still re-run the range
optimizer, but we save the quick select we already have, and take the
best of the two quick selects.